Correcciones de Description Helper
