+++ app/src/main/java/ayds/newyork/songinfo/home/model/repository/external/spotify/tracks/SpotifyToSongResolver.kt
@@ -21,6 +21,9 @@ private const val RELEASE_DATE_PRECISION = "release_date_precision"
 private const val URL = "url"
 private const val EXTERNAL_URL = "external_urls"
 private const val SPOTIFY = "spotify"
+private const val SPOTIFY_DAY = "day"
+private const val SPOTIFY_MONTH = "month"
+private const val SPOTIFY_YEAR = "year"
 
 internal class JsonToSongResolver : SpotifyToSongResolver {
 
@@ -70,7 +73,12 @@ internal class JsonToSongResolver : SpotifyToSongResolver {
 
     private fun JsonObject.getReleaseDatePrecision(): DatePrecision {
         val album = this[ALBUM].asJsonObject
-        return DatePrecision.valueOf(album[RELEASE_DATE_PRECISION].asString.uppercase())
+        return when(album[RELEASE_DATE_PRECISION].asString){
+            SPOTIFY_DAY -> DatePrecision.DAY
+            SPOTIFY_MONTH -> DatePrecision.MONTH
+            SPOTIFY_YEAR -> DatePrecision.YEAR
+            else -> DatePrecision.EMPTY
+        }
     }
 
     private fun JsonObject.getImageUrl(): String {
