+++ app/src/main/java/ayds/newyork/songinfo/home/view/HomeViewInjector.kt
@@ -5,7 +5,7 @@ import ayds.newyork.songinfo.home.model.HomeModelInjector
 
 object HomeViewInjector {
 
-    val songDescriptionHelper: SongDescriptionHelper = SongDescriptionHelperImpl(DateFormatter())
+    val songDescriptionHelper: SongDescriptionHelper = SongDescriptionHelperImpl(DateFormatterImpl())
 
     fun init(homeView: HomeView) {
         HomeModelInjector.initHomeModel(homeView)
